Update patch set 3
Patch Set 3:

Can you rebase this on top of the merge commit?  That merge needs to keep the same sha1 as referenced in the core gerrit merge commit. I'm not sure if it will work if this commit is merged first.

Patch-set: 3
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>