Update patch set 2
Patch Set 2:

I would do the merge first, so that the test is in the repo, and then the fix as a separate commit on top of it.  Both commits should be set with the same topic so that they get submitted at the same time.

Patch-set: 2
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>