Grant the release managers the permissions to push a signed tag
Allow the automation of the release process from the CI, where we use an
automated job with a reviewed script to generate and tag the release of
core plugins, instead of a manual operation from the release manager.

The use of non-interactive GPG keys and a totally automated and
scripted process helps prevent a future supply chain attack where a
malicious user tries to eavesdrop a maintainer's credentials using a
phishing attack.


Change-Id: Ic915da2dbe82397e42d60a55ca4734fa9a4422ce